drop support for char8_t and float32_t, add support for signed and unsigned 8, 16, 32, and 64 bit integers. add support for bin and hex literals and respective semantic checks.

+++ programs/sieve.cpp
@@ -1,11 +1,11 @@
 #include "FrankoRuntime.hpp"
 
 int main() {
-    int32_t n;
+    uint32_t n;
     n = 80;
-    Franko_Dynamic_Array<int32_t> isPrime;
+    Franko_Dynamic_Array<uint8_t> isPrime;
     isPrime.init((n + 1));
-    int32_t i;
+    uint32_t i;
     i = 0;
     while ((i <= n))
     {
@@ -14,13 +14,13 @@ int main() {
     }
     isPrime[0] = 0;
     isPrime[1] = 0;
-    int32_t p;
+    uint32_t p;
     p = 2;
     while (((p * p) <= n))
     {
         if ((isPrime[p] != 0))
         {
-            int32_t j;
+            uint32_t j;
             j = (p * p);
             while ((j <= n))
             {
